requirements.txt
matplotlib==3.10.3
pandas==2.3.0
numpy==1.26.4
seaborn==0.13.2
scipy==1.15.3
scikit-learn==1.7.0
nltk==3.9.1
scikit-surprise==1.1.4
pytest==8.4.1
Flask==3.1.1